ext/POSIX/t/sysconf.t: only unlink the fifo if we have a name for it
This warned on Win32 since the whole test file is skipped due
to the lack of anything sysconf() to test.

+++ ext/POSIX/t/sysconf.t
@@ -158,7 +158,9 @@ SKIP: {
 }
 
 END {
-    1 while unlink($fifo);
+    if ($fifo) {
+        1 while unlink($fifo);
+    }
 }
 
 SKIP: {
